requirements.txt
django
django-restframework
django-rest-authemail
requests